+++ Sources/NeuralNetwork/NeuralNetwork.c
@@ -64,7 +64,7 @@ struct Network init(int nb_layer, int *nb_neuron)
 }
 
 
-char execute_network(struct Network *network, int *inputs,int index_val, double *value, Game *game)
+char execute_network(struct Network *network, double *inputs,int index_val, double *value, Game *game)
 {
 
   //enter the input
